requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
llama-index-core==0.10.13
llama-index-embeddings-huggingface==0.1.4
llama-index-llms-llama-cpp==0.1.3
llama-index-llms-ollama==0.1.2
llama-index-readers-file==0.1.6
llama-index-vector-stores-postgres==0.1.2
llama_cpp_python==0.2.53
llamaindex-py-client==0.1.13
MarkupSafe==2.1.5
marshmallow==3.21.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.1
pgvector==0.2.5
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.6.2
pydantic_core==2.16.3
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pypdf==4.0.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.38.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
